backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
supabase>=2.5.1
google-genai>=1.0.0
youtube-transcript-api>=0.6.2
pymupdf>=1.24.7
python-dotenv>=1.0.1
pydantic>=2.8.2
tiktoken>=0.7.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9
